+++ activerecord/lib/active_record/model_schema/schema_context.rb
@@ -18,8 +18,10 @@ module ActiveRecord
       end
 
       def attributes_builder
-        defaults = _default_attributes.except(*(column_names - Array(primary_key)))
-        ActiveModel::AttributeSet::Builder.new(attribute_types, defaults)
+        @attributes_builder ||= begin
+          defaults = _default_attributes.except(*(column_names - Array(primary_key)))
+          ActiveModel::AttributeSet::Builder.new(attribute_types, defaults)
+        end
       end
 
       def columns_hash
@@ -47,7 +49,7 @@ module ActiveRecord
       end
 
       def column_names
-        columns.map(&:name).freeze
+        @column_names ||= columns.map(&:name).freeze
       end
 
       def symbol_column_to_string(name_symbol)
@@ -115,9 +117,11 @@ module ActiveRecord
         @symbol_column_to_string_name_hash = nil
         @content_columns = nil
         @column_defaults = nil
+        @column_names = nil
         @columns = nil
         @columns_hash = nil
         @schema_loaded = false
+        @attributes_builder = nil
         @attribute_types = nil
         @default_attributes = nil
       end